refact: change folders and add validation
