+++ backend/src/main/java/com/techmaki/sushi_hub/orders/application/dtos/CreateOrderItemRequest.java
@@ -1,0 +1,22 @@
+package com.techmaki.sushi_hub.orders.application.dtos;
+
+import jakarta.validation.constraints.NotNull;
+
+public record CreateOrderItemRequest(
+
+    @NotNull
+    long productId,
+
+    @NotNull
+    int quantity,
+
+    @NotNull
+    double price,
+
+    double discount,
+
+    @NotNull
+    double totalPrice
+) {
+
+}
