+++ backend/src/main/java/com/techmaki/sushi_hub/orders/application/dtos/CreateOrderRequest.java
@@ -1,9 +1,14 @@
-package com.techmaki.sushi_hub.orders.domain.dtos;
+package com.techmaki.sushi_hub.orders.application.dtos;
 
 import java.util.List;
 
+import jakarta.validation.constraints.NotNull;
+
 public record CreateOrderRequest(
+    
+    @NotNull
     Long userId,
+
     List<CreateOrderItemRequest> items
 ) {
