+++ backend/src/main/java/com/techmaki/sushi_hub/orders/application/dtos/UpdateOrderItemRequest.java
@@ -1,4 +1,4 @@
-package com.techmaki.sushi_hub.orders.domain.dtos;
+package com.techmaki.sushi_hub.orders.application.dtos;
 
 public record UpdateOrderItemRequest(
     Long orderId,
