--- backend/src/main/java/com/techmaki/sushi_hub/orders/domain/dtos/CreateOrderItemRequest.java
@@ -1,12 +1,0 @@
-package com.techmaki.sushi_hub.orders.domain.dtos;
-
-public record CreateOrderItemRequest(
-
-    long productId,
-    int quantity,
-    double price,
-    double discount,
-    double totalPrice
-) {
-
-}
